separated spec files for container
