+++ cypress/integration/v2/blocks/container/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Container Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
