+++ cypress/integration/v2/blocks/container/mobile-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileAdvanced } from '.'
+
+describe( 'Container Block ( Mobile Advanced )', registerTests( [ mobileAdvanced ] ) )
