+++ cypress/integration/v2/blocks/container/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Container Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
